refactor to query.ts
